tests/wpt/tests/css/css-animations/parsing/animation-trigger-type-invalid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-trigger-type with invalid values</title>
<!-- TODO(crbug.com/390314945): Replace with spec link when the spec lands. -->
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8942 ">
<meta name="assert" content="animation-trigger-type supports only the grammar '<single-animation-trigger-type> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-trigger-type", "auto");
test_invalid_value("animation-trigger-type", "once repeat");

test_invalid_value("animation-trigger-type", "once, initial");
test_invalid_value("animation-trigger-type", "initial, once");
</script>
</body>
</html>
